requirements.txt
streamlit
pandas
openpyxl
sentence-transformers
faiss-cpu
